// src/types/tableTypes.ts
export interface Operator {
    operator_id: string;         // Уникальный идентификатор оператора
    phone: string;         // Уникальный идентификатор оператора
    tg_account: string;        // ID оператора
    created_at: string;          // Дата добавления оператора
    status: operator_status_enum;               // Статус пользователя
}
// ...
export type operator_status_enum ='free'| 'busy'